+++ upkie/envs/upkie_base_env.py
@@ -169,9 +169,7 @@ class UpkieBaseEnv(abc.ABC, gymnasium.Env):
         self._spine.stop()
         self.__reset_rate()
         self.__reset_init_state()
-        self._spine.start(self._spine_config)
-        self._spine.get_observation()  # might be a pre-reset observation
-        spine_observation = self._spine.get_observation()
+        spine_observation = self._spine.start(self._spine_config)
         self.parse_first_observation(spine_observation)
         observation = self.get_env_observation(spine_observation)
         info = {"spine_observation": spine_observation}
@@ -232,7 +230,7 @@ class UpkieBaseEnv(abc.ABC, gymnasium.Env):
             self.__rate.sleep()  # wait until clock tick to send the action
             self.log("rate", {"slack": self.__rate.slack})
 
-        # Act
+        # Prepare spine action
         spine_action = self.get_spine_action(action)
         for key in ("bullet", "log"):
             if not self.__extras[key]:
@@ -240,10 +238,11 @@ class UpkieBaseEnv(abc.ABC, gymnasium.Env):
             spine_action[key] = {}
             spine_action[key].update(self.__extras[key])
             self.__extras[key].clear()
-        self._spine.set_action(spine_action)
 
-        # Observe
-        spine_observation = self._spine.get_observation()
+        # Send action to and get observation from the spine
+        spine_observation = self._spine.set_action(spine_action)
+
+        # Process spine observation
         observation = self.get_env_observation(spine_observation)
         reward = self.get_reward(observation, action)
         terminated = self.detect_fall(spine_observation)